LayoutTests/platform/wpe/accessibility/scroll-to-make-visible-iframe-offscreen-expected.txt

This tests that scrolling to make an element visible works properly when there's an iframe off screen.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


5000-pixel box


Test scrolling an offscreen iframe.
PASS window.pageYOffset is 0

Scroll lower target to visible.
The iframe should be scrolled into view
FAIL scrolledYOffset > 0 should be true. Was false.
Test the lower target should be scrolled into view.
FAIL scrolledIntoView should be true. Was false.

Scroll upper target to visible.
The main window shouldn't scroll.
PASS window.pageYOffset == scrolledYOffset is true
Test the upper target should be scrolled into view.
PASS scrolledIntoView is true

Reset scrolling. Test scrolling in nested iframes.
PASS window.pageYOffset is 0

Scroll inner button to visible.
Test the button inside the inner frame should be scrolled into view
FAIL scrolledIntoView should be true. Was false.
The inner iframe should be scrolled into view
FAIL outterFrameWindow.pageYOffset > 0 should be true. Was false.

Scroll outter text element to visible.
The Y offset of the outter iframe should be reset.
PASS outterFrameWindow.pageYOffset is 0
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
